+++ research/utils/engine_logging.py
@@ -1,48 +1,10 @@
+import pdb
 from enum import Enum
 from typing import Callable, List, Dict
 from ignite.engine import Engine
 from torch.utils.data import DataLoader
-
-
-def _to_stdout(engine: Engine, fields: List[str], epoch_num=None) -> None:
-    """Prints string formatted engine output fields to stdout"""
-    log_str = "  ".join(
-        ["{}: {:.3f}".format(field, engine.state.output[field]) for field in fields]
-    )
-    log_str = "Epoch[{:d}] {}".format(tern(epoch_num, engine.state.epoch), log_str)
-    print(log_str)
-
-
-def _to_log(engine: Engine, fields: List[str], epoch_num=None) -> None:
-    """Logs string formatted engine output fields to logfile"""
-    log_str = "  ".join(
-        ["{}: {:.3f}".format(field, engine.state.output[field]) for field in fields]
-    )
-    log_str = "Epoch[{:d}] {}".format(tern(epoch_num, engine.state.epoch), log_str)
-    engine.logger.info(log_str)
-
-
-def _to_file(engine: Engine, fields: List[str]) -> None:
-    """Logs formatted engine output fields to separate file (binary)"""
-    for field in fields:
-        # Check that engine has fp in state
-        # Check
-        pass
-
-
-class LOG_MODE(Enum):
-    # Log to standard out (print())
-    STDOUT = _to_stdout
-    # Log as message in log file
-    LOG_MSG = _to_log
-    # Log to separate file
-    LOG_FIL = _to_file
-    # Log image to standalone folder
-    LOG_IMG = 3
-    # Log to visdom
-    VISDOM_NUM = 4
-    # Log image to visdom
-    VISDOM_IMG = 5
+from utils.log_modes import LOG_MODE
+from utils.helpers import tern
 
 
 def _lf(
@@ -67,22 +29,13 @@ def _lf_val(
     )
 
 
-# FIXME: Engine doesn't seem to have "times" attribute in engine state contrary to docs
-# def log_total_time(engine: Engine) -> None:
-#     """Log the total time to complete training"""
-#     engine.logger.info("Total: {}".format(engine.state.times["COMPLETED"]))
-
-
 def log_engine_output(
     engine: Engine, fields: Dict[LOG_MODE, List[str]], epoch_num=None
 ) -> None:
     """Log numerical fields in the engine output dictionary to stdout"""
-    import pdb
-
     pdb.set_trace()
     for mode in fields.keys():
-        # mode((engine, fields[mode], epoch_num)) confirm this
-        pass
+        mode(engine, fields[mode], "output")
 
 
 def log_engine_metrics(
@@ -91,16 +44,14 @@ def log_engine_metrics(
 
     """Run engine on Dataloader. Then log numerical fields in the engine metrics dictionary to stdout"""
     engine.run(loader)
-    metrics = engine.state.metrics
-    log_str = "  ".join(
-        ["{}: {:.3f}".format(field, engine.state.metrics[field]) for field in fields]
-    )
-    log_str = "Epoch[{:d}] {}".format(tern(epoch_num, engine.state.epoch), log_str)
-    engine.logger.info(log_str)
-
-    if stdout:
-        print(log_str)
+    for mode in fields.keys():
+        mode(engine, fields[mode], "metrics")
 
 
 # TODO: Add Visdom logging callbacks
 # TODO: Add Slack notification callback
+
+# FIXME: Engine doesn't seem to have "times" attribute in engine state contrary to docs
+# def log_total_time(engine: Engine) -> None:
+#     """Log the total time to complete training"""
+#     engine.logger.info("Total: {}".format(engine.state.times["COMPLETED"]))
